+++ app/(home)/JournalAvatarUpload.tsx
@@ -2,7 +2,7 @@ import { IconPhotoPlus } from "@tabler/icons-react";
 import { UseFormRegister } from "react-hook-form";
 import React, { useState } from "react";
 import Image from "next/image";
-import { cn } from "../lib/cn";
+import { cn } from "../assets/lib/cn";
 
 type JournalAvatarUploadProps = {
   register: UseFormRegister<any>;
