+++ app/(layout)/Navbar.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import ChainSheetsIcon from "../icons/ChainSheetsIcon";
+import ChainSheetsIcon from "../assets/icons/ChainSheetsIcon";
 import { IconCirclePlus, IconWallet } from "@tabler/icons-react";
 import Link from "next/link";
-import { button } from "../lib/variants";
+import { button } from "../assets/lib/variants";
 
 const Navbar = () => {
   return (
