+++ app/assets/icons/ChainSheetsIcon.tsx
