+++ app/assets/lib/shortenAddress.ts
