+++ app/assets/lib/variants.tsx
